# packages/api-spec/openapi.yaml
openapi: 3.0.3
info:
  title: Datememe API
  version: 0.0.1
  description: >
    MVP contract for the favorites-based match-making app. See
    /docs/data-schema-proposal.md and /docs/project-requirements.md for the
    product and data-model rationale behind these routes.
  license:
    name: UNLICENSED
servers:
  - url: http://localhost:3001
    description: Local dev
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >
        Cookie-first (web), Bearer-header fallback (native apps via
        expo-secure-store). See packages/sdk/src/client.ts.
    adminAuth:
      type: http
      scheme: bearer
      description: >
        Same token as bearerAuth, plus a role check (User.role === ADMIN) —
        see plugins/security.ts's adminAuth. Used on operations restricted to
        the DATEMEME admin area.
  parameters:
    Cursor:
      in: query
      name: cursor
      schema: { type: string }
      description: Opaque cursor returned by the previous page.
    Limit:
      in: query
      name: limit
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
...
    GroupSlugs:
      in: query
      name: groupSlugs
      description: >
        Comma-separated CategoryGroup slugs (e.g. "music,food") — multi-select.
        OR'd together: a real, server-side filter to *any* selected group,
        not an AND across all of them. Comma-joined rather than a repeated
        query key so a single selection never needs array coercion on either
        the client or the server's querystring parser. Shared by GET
        /lists/feed and GET /discover/feed — the two pages' category chips
        run the same filtering logic, though what a "match" means differs
        per feed: on /lists/feed it's literal category membership; on
        /discover/feed it's the *taste* axis — real engagement (a started or
        completed list) in that group, not people tagged with it.
      schema: { type: string }
  responses:
    Unauthorized:
      description: Not authenticated
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: Not permitted
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
    PaginatedMeta:
      type: object
      required: [hasMore, nextCursor]
      properties:
        hasMore: { type: boolean }
        nextCursor: { type: string, nullable: true }
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }
        code: { type: string }

    # ── Identity ──────────────────────────────────────────
    Profile:
      type: object
      required:
        [
          id,
          userId,
          username,
          displayName,
          genderIdentity,
          bio,
          seekingGenders,
          locationLabel,
          avatarUrl,
          photos,
          photosLocked,
          isDiscoverable,
          onboardingStep,
        ]
      properties:
        id: { type: string }
        userId: { type: string }
        username: { type: string }
        displayName: { type: string }
        genderIdentity: { type: string, nullable: true }
        bio:
          type: string
          nullable: true
          maxLength: 150
          description: Short optional tagline — not a bio essay, see docs §1/§3.
        seekingGenders:
          type: array
          items: { type: string }
        locationLabel: { type: string, nullable: true }
        avatarUrl:
          type: string
          nullable: true
          description: Null for a free-tier viewer looking at someone else's profile — see docs §8 photo gate. Also null when the profile genuinely has no avatar; see `photosLocked` to tell the two apart.
        photos:
          type: array
          items: { type: string }
          description: Gallery beyond the primary avatarUrl; empty for a free-tier viewer looking at someone else's profile (docs §8), and also empty when the profile genuinely has no gallery photos.
        photosLocked:
          type: boolean
          description: >
            True only when `avatarUrl`/`photos` were withheld by the docs §8
            photo gate (a free-tier viewer looking at someone else's
            profile) — always false for your own profile, and false for a
            profile that simply has no photos, so a client can tell "go
            premium to see this" apart from "there's nothing to see."
        isDiscoverable: { type: boolean }
        onboardingStep: { type: integer }
...
    ImageCredit:
      type: object
      nullable: true
      properties:
        attribution: { type: string, nullable: true }
        landingUrl: { type: string, nullable: true }
        licenseUrl: { type: string, nullable: true }
        provider: { type: string, nullable: true }
...
    Entity:
      type: object
      required: [id, entityTypeId, canonicalName, slug, imageUrl, metadata, status, usageCount]
      properties:
        id: { type: string }
        entityTypeId: { type: string }
        canonicalName: { type: string }
        slug: { type: string }
        imageUrl: { type: string, nullable: true }
        imageCredit: { $ref: '#/components/schemas/ImageCredit' }
        metadata: { type: object, nullable: true }
        status: { type: string, enum: [PENDING, APPROVED, REJECTED, MERGED] }
        usageCount: { type: integer }
...
    SharedFavorite:
      type: object
      required: [categoryShortLabel, entityName]
      properties:
        categoryShortLabel: { type: string }
        entityName: { type: string }
    MatchInsight:
      type: object
      required: [icon, title, description]
      properties:
        icon:
          type: string
          description: >
            A short symbolic key (e.g. "shared-taste", "rare-overlap") — the
            client owns the actual glyph/emoji.
        title: { type: string }
        description: { type: string }
...
    AgeBucket:
      type: string
      enum: [20s, 30s, 40s, 50plus]

    # ── Quick Picks (taste-graph preference signals) ─────────
...
    MetricType:
      type: string
      enum: [popularity, community-position, viewer-position, percentile, overlap, rarity, trend, completion]
    Metric:
      type: object
      required: [type, label, value]
      properties:
        type: { $ref: '#/components/schemas/MetricType' }
        label: { type: string }
        value:
          description: Already display-formatted ("82%", "3.4×") or a raw number — renderer's choice.
          oneOf: [{ type: string }, { type: number }]
        importance:
          type: string
          enum: [primary, secondary, tertiary]
          description: >
            Render-priority hint, not a domain fact — structures pick which
            metrics to show under their render budget (proposal §5) using
            this field; it never changes what the metric means.
    Capabilities:
      type: object
      properties:
        canFavorite: { type: boolean }
        canRank: { type: boolean }
        canMessage: { type: boolean }
        canOpenProfile: { type: boolean }
    ContentUnitRelationship:
      type: object
      properties:
        favorited: { type: boolean }
        completed: { type: boolean }
    ContentUnitKind:
      type: string
      enum: [poll, person, category, entity, insight]
    ContentUnit:
      type: object
      required: [id, kind, title]
      properties:
        id:
          type: string
          description: Stable within its collection. For kind='category' this is the Category slug (routable); for kind='person' the Profile id.
        kind: { $ref: '#/components/schemas/ContentUnitKind' }
        title: { type: string }
        subtitle: { type: string, nullable: true }
        imageUrl: { type: string, nullable: true }
        imageCredit: { $ref: '#/components/schemas/ImageCredit' }
        metrics:
          type: array
          items: { $ref: '#/components/schemas/Metric' }
        capabilities: { $ref: '#/components/schemas/Capabilities' }
        position:
          type: integer
          nullable: true
          description: >
            Order of this unit *within its collection* only — not a
            comparative rank. Comparative standing is a fact about the unit
            and belongs under `metrics` as community-position /
            viewer-position / percentile (proposal §4).
        relationship: { $ref: '#/components/schemas/ContentUnitRelationship' }
        entity:
          description: Present for kind='category' (the top pick, if any) and kind='entity'.
          type: object
          nullable: true
          allOf: [{ $ref: '#/components/schemas/Entity' }]
        profile:
          description: Present for kind='person' — the full candidate, so the renderer needs no second fetch to navigate to ProfileDetail.
          type: object
          nullable: true
          allOf: [{ $ref: '#/components/schemas/Profile' }]
        age:
          description: Present for kind='person' only — mirrors DiscoveryCandidate.age.
          type: integer
        sharedFavorites:
          description: >
            Present for kind='person' only — the actual overlapping picks
            ("You both ranked"), not just a count. This is the taste content
            that explains the match; render it more prominently than
            `alsoInto` (proposal correction: shared taste is the reason the
            match is relevant, independent interests are supporting texture).
          type: array
          items: { $ref: '#/components/schemas/SharedFavorite' }
        alsoInto:
          description: >
            Present for kind='person' only — a few of the candidate's own
            top-ranked picks outside the overlap with the viewer ("Also
            into"). Supporting texture, not the headline.
          type: array
          items: { $ref: '#/components/schemas/Entity' }
        insights:
          description: Present for kind='person' only — mirrors DiscoveryCandidate.insights.
          type: array
          items: { $ref: '#/components/schemas/MatchInsight' }
        previewEntities:
          description: >
            Top-ranked entities for a completed list (kind='category' with
            relationship.completed=true) — lets the Grid's completed-list
            variant show a multi-item preview instead of one thumbnail.
          type: array
          items: { $ref: '#/components/schemas/Entity' }
    CollectionType:
      type: string
      enum: [favorites, lists, recommendations, comparison, trending, prompt, quiz]
    ContentCollectionContext:
      type: object
      properties:
        reason: { type: string }
        sourceEntityId: { type: string }
        sourceEntityType: { type: string }
        groupSlug:
          type: string
          description: >
            Present on a CategoryGroup topic module (getListsFeed's
            groupModules, including continuation chunks) — the chip id it
            belongs to. Lets the client's chip bar reorder that module to the
            top of the feed instead of scrolling to find it, without
            string-matching module ids.

    # ContentCollection and InteractiveModule are documented here as the two
    # logical shapes a FeedModule takes, but the wire schema below (FeedModule)
    # is deliberately ONE flat object rather than a oneOf of these two: fast-json-stringify
    # (Fastify's response serializer) can't reliably disambiguate a oneOf of
    # structurally-similar objects — it needs exactly one branch to match and
    # throws "does not match schema definition" otherwise. moduleKind is what
    # actually discriminates the two on the wire; TypeScript narrows on it
    # client-side (see apps/mobile/src/ui/content/types.ts).
    FeedModule:
      type: object
      required: [moduleKind, id]
      properties:
        moduleKind: { type: string, enum: [collection, interactive] }
        id: { type: string }
        # Populated when moduleKind='collection' (a ContentCollection):
        type: { $ref: '#/components/schemas/CollectionType' }
        title: { type: string, nullable: true }
        context: { $ref: '#/components/schemas/ContentCollectionContext' }
        suggestedStructure:
          type: string
          enum: [grid, rail, spotlight, river]
          description: Advisory only — the frontend has final say per viewport/context (proposal §1).
        options:
          type: object
          properties:
            gridShape: { type: string, enum: [square, dense] }
            columns:
              type: integer
              description: >
                Advisory column count for a Grid structure at desktop width
                (mobile always collapses to 1) — lets the backend request
                "large, few-per-row" cards without a rigid gridShape enum.
        items:
          type: array
          items: { $ref: '#/components/schemas/ContentUnit' }
        # Populated when moduleKind='interactive' (an InteractiveModule):
        kind: { type: string, enum: [quick-picks] }
    SummaryStat:
      type: object
      required: [label, value]
      properties:
        label: { type: string }
        value:
          oneOf: [{ type: string }, { type: number }]
    PageSummary:
      type: object
      required: [title]
      properties:
        title: { type: string }
        subtitle: { type: string, nullable: true }
        stats:
          type: array
          items: { $ref: '#/components/schemas/SummaryStat' }
        featuredEntity:
          type: object
          nullable: true
          allOf: [{ $ref: '#/components/schemas/Entity' }]
    FilterChip:
      type: object
      required: [id, label]
      properties:
        id: { type: string }
        label: { type: string }
    ContentFeedPageResponse:
      type: object
      required: [data, meta]
      properties:
        filterNotice: { type: string }
        summary:
          description: Page chrome — only ever populated on the first page (cursor omitted).
          allOf: [{ $ref: '#/components/schemas/PageSummary' }]
        chips:
          type: array
          items: { $ref: '#/components/schemas/FilterChip' }
        data:
          type: array
          items: { $ref: '#/components/schemas/FeedModule' }
        meta: { $ref: '#/components/schemas/PaginatedMeta' }
    ContentUnitPageResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/ContentUnit' }
        meta: { $ref: '#/components/schemas/PaginatedMeta' }

    # ── Admin ───────────────────────────────────────────────
    # apps/admin (Vite) -> apps/mobile Admin migration — see CLAUDE.md.
    # Migrated so far: Dashboard (/admin/queue), Moderation (this batch).
    # The remaining ~20 /admin/* routes stay hand-registered on the Fastify
    # instance until their own migration batch.
...
    AdminCategoryScalar:
      type: object
      required:
        [
          id,
          groupId,
          entityTypeId,
          parentEntityId,
          slug,
          prompt,
          shortLabel,
          minItems,
          maxItems,
          orderingMode,
          isMatchSignal,
          isPremiumOnly,
          isActive,
        ]
      properties:
        id: { type: string }
        groupId: { type: string }
        entityTypeId: { type: string }
        parentEntityId: { type: string, nullable: true }
        slug: { type: string }
        prompt: { type: string }
        shortLabel: { type: string }
        minItems: { type: integer }
        maxItems: { type: integer }
        orderingMode: { type: string, enum: [RANKED, UNRANKED] }
        isMatchSignal: { type: boolean }
        isPremiumOnly: { type: boolean }
        isActive: { type: boolean }
...
    AdminSitePickGroupScalar:
      type: object
      required: [id, slug, label, sortOrder, isActive]
      properties:
        id: { type: string }
        slug: { type: string }
        label: { type: string }
        sortOrder: { type: integer }
        isActive: { type: boolean }
    AdminSitePickItem:
      description: One List Definition (Category row) curated into a Site Picks group, in display order. Has no `id` of its own — SitePickItem is a (groupId, categoryId) composite key in the schema.
      type: object
      required: [groupId, categoryId, sortOrder, category]
      properties:
        groupId: { type: string }
        categoryId: { type: string }
        sortOrder: { type: integer }
        category: { $ref: '#/components/schemas/AdminCategoryScalar' }
    AdminSitePickGroup:
      description: A Site Picks group as seen by the admin browser/editor — the group plus its curated List Definitions, in order.
      allOf:
        - $ref: '#/components/schemas/AdminSitePickGroupScalar'
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items: { $ref: '#/components/schemas/AdminSitePickItem' }
    AdminCreateSitePickGroupInput:
      type: object
      required: [slug, label]
      properties:
        slug: { type: string }
        label: { type: string }
        sortOrder: { type: integer }
        isActive: { type: boolean }
    AdminUpdateSitePickGroupInput:
      type: object
      required: [slug, label]
      properties:
        slug: { type: string }
        label: { type: string }
        sortOrder: { type: integer }
        isActive: { type: boolean }
    AdminSitePickItemInput:
      type: object
      required: [categoryId, sortOrder]
      properties:
        categoryId: { type: string }
        sortOrder: { type: integer }
    AdminUpdateSitePickItemsInput:
      type: object
      required: [items]
      properties:
        items:
          description: Full replace, not a merge. Intended to be exactly 4 entries — enforced by the admin UI, not the server.
          type: array
          items: { $ref: '#/components/schemas/AdminSitePickItemInput' }
...
paths:
...
  /lists/feed:
    get:
      operationId: getListsFeed
      summary: The Lists page as a PageSummary + FeedModule stream, grouped by topic
      tags: [content-feed]
      description: >
        Backend adapter over the existing Lists/Taxonomy data (TaxonomyService,
        ListService), reshaped per docs/shared-content-system-proposal.md. One
        ContentCollection per non-empty CategoryGroup ("Music", "Movies & TV",
        ...), each mixing answered and unanswered categories — completion is
        a fact carried by the unit (`relationship.completed`), not a separate
        section. `chips` mirror the same groups; with no `groupSlugs`, the
        whole feed loads unfiltered ("Your lists" -> Site Picks -> every
        topic). With `groupSlugs`, this is a real filtered query — "Your
        lists" still leads, then only the selected group(s)' Site Picks and
        topic sections, with Quick Picks folded in after them — same
        filtering shape as GET /discover/feed.
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/GroupSlugs'
      responses:
        '200':
          description: Page summary, chips, and one feed module per topic
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ContentFeedPageResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /lists/feed/collections/{collectionId}:
    get:
      operationId: getListsFeedCollection
      summary: Paginate a single Lists-feed collection ("View all a topic's lists")
      tags: [content-feed]
      description: collectionId is a CategoryGroup slug (e.g. "music") — see GET /lists/feed.
      parameters:
        - in: path
          name: collectionId
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Units within the collection
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ContentUnitPageResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /discover/feed:
    get:
      operationId: getDiscoverFeed
      summary: The Discover page as a People-grid feed with contextual interruptions
      tags: [content-feed]
      description: >
        Wraps the existing rarity-weighted overlap feed (GET /discovery) as
        the baseline Grid, adding a "similar taste" Rail keyed off the
        viewer's own #1 pick, one embedded Quick Picks interactive module,
        and a conditional Spotlight for an exceptional match — all
        optional/sparse per docs §1. Mutual gender compatibility is always
        applied as a baseline correctness rule, not an optional filter.
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/GroupSlugs'
        - in: query
          name: nearMe
          description: Requires the viewer to have their own location set — returns nobody outside the radius rather than silently ignoring the filter.
          schema: { type: boolean }
        - in: query
          name: ageBucket
          schema: { $ref: '#/components/schemas/AgeBucket' }
      responses:
        '200':
          description: Feed modules (no PageSummary — Discover has no viewer-aggregate header)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ContentFeedPageResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /discover/feed/collections/{collectionId}:
    get:
      operationId: getDiscoverFeedCollection
      summary: Paginate a single Discover-feed collection ("View all")
      tags: [content-feed]
      parameters:
        - in: path
          name: collectionId
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Units within the collection
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ContentUnitPageResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ── Quick Picks ─────────────────────────────────────────
...
  /admin/site-picks/groups:
    get:
      operationId: getSitePickGroups
      summary: List every Site Picks group, with its curated List Definitions, in order
      tags: [admin]
      security:
        - adminAuth: []
      responses:
        '200':
          description: All Site Picks groups
          content:
            application/json:
              schema:
                type: object
                required: [groups]
                properties:
                  groups:
                    type: array
                    items: { $ref: '#/components/schemas/AdminSitePickGroup' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      operationId: createSitePickGroup
      summary: Create a new Site Picks group
      tags: [admin]
      security:
        - adminAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AdminCreateSitePickGroupInput' }
      responses:
        '201':
          description: The created group
          content:
            application/json:
              schema:
                type: object
                required: [group]
                properties:
                  group: { $ref: '#/components/schemas/AdminSitePickGroupScalar' }
        '400':
          description: Missing field
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /admin/site-picks/groups/{id}:
    put:
      operationId: updateSitePickGroup
      summary: Update a Site Picks group's label, slug, order, or active state
      tags: [admin]
      security:
        - adminAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AdminUpdateSitePickGroupInput' }
      responses:
        '200':
          description: The updated group
          content:
            application/json:
              schema:
                type: object
                required: [group]
                properties:
                  group: { $ref: '#/components/schemas/AdminSitePickGroupScalar' }
        '400':
          description: Missing field
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      operationId: deleteSitePickGroup
      summary: Delete a Site Picks group (does not touch the underlying List Definitions)
      tags: [admin]
      security:
        - adminAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '204':
          description: Deleted
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /admin/site-picks/groups/{id}/items:
    put:
      operationId: updateSitePickGroupItems
      summary: Replace this Site Picks group's curated List Definitions and their order (full replace, not a merge)
      tags: [admin]
      security:
        - adminAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AdminUpdateSitePickItemsInput' }
      responses:
        '200':
          description: The group's curated List Definitions after the replace
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/AdminSitePickItem' }
        '400':
          description: items must be an array
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ── Admin: Media (batch 7) ────────────────────────────────
...